+++ vms/UserListPage.php
@@ -60,7 +60,20 @@ class UserListPage {
                     <td><?= $row['name_commune'] ?></td>
                     <td><?= $row['status'] ?></td>
                     <td><?= $row['active'] ?></td>
-                    <td class="center"><i class="fas fa-trash"></i> <a href="/deleteuser/<?= $row['id'] ?>"> Delete</a></td>
+                    <td class="center"><i class="fas fa-trash"></i> <a data-href="/deleteuser/<?= $row['id'] ?>" data-target="#confirm-delete" data-toggle="modal"> Delete</a></td>
+                    <div class="modal fade" id="confirm-delete" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+                        <div class="modal-dialog">
+                            <div class="modal-content">
+                                <div class="modal-body">
+                                    Bạn có chắc chắn sẽ muốn xóa người dùng này ?
+                                </div>
+                                <div class="modal-footer">
+                                    <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
+                                    <a class="btn btn-danger btn-ok">Delete</a>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
                     <td class="center"><i class="fas fa-edit"></i> <a href="/edituser/<?= $row['id'] ?>">Edit</a></td>
                     <td class="center"><i class="fas fa-lock"></i> <a href="/blockuser/<?= $row['id'] ?>">Block</a></td>
                     <td class="center"><i class="fas fa-check-circle"></i> <a href="/enableuser/<?= $row['id'] ?>">Enable</a></td>
